rename `tailwind` to `base` in components
