+++ packages/refine-vue/src/components/base/clauses/numeric/base-numeric-single-clause.vue
@@ -6,7 +6,7 @@
 
 <script>
 import { numericSingleProps } from './mixins';
-import NumberInput from '../../../../components/tailwind/inputs/number-input';
+import NumberInput from '../../../../components/base/inputs/number-input';
 
 export default {
   name: 'base-numeric-single-clause',
