+++ packages/refine-vue/src/components/base/conditions/base-condition.vue
@@ -13,7 +13,7 @@
 </template>
 
 <script>
-import Selector from '../../../components/tailwind/selector';
+import Selector from '../../../components/base/selector';
 
 export default {
   name: 'base-condition',
