+++ packages/refine-vue/src/components/base/conditions/numeric-condition.vue
@@ -20,7 +20,7 @@
 import BaseCondition from './base-condition';
 import { conditionProps } from '../../../mixins';
 import { NumericCondition as RenderlessNumericCondition } from '../../../components/renderless/conditions';
-import * as numericClauseOptions from '../../../components/tailwind/clause-options/numeric';
+import * as numericClauseOptions from '../../../components/base/clause-options/numeric';
 
 export default {
   name: 'numeric-condition',
