+++ packages/refine-vue/src/components/base/index.js
@@ -1,6 +1,6 @@
 import Selector, { SelectorOption } from './selector';
 import ConditionSelector from './condition-selector';
 import Query from '../../components/renderless/query.vue';
-import QueryBuilder from '../../components/tailwind/query-builder';
+import QueryBuilder from '../../components/base/query-builder';
 
 export { ConditionSelector, Query, QueryBuilder, Selector, SelectorOption };
