+++ packages/refine-vue/src/compositions/tailwind/useNumericDoubleClause.js
@@ -1,5 +1,5 @@
 import { h } from 'vue-demi';
-import BaseNumericDoubleClause from '../../components/tailwind/clauses/numeric/base-numeric-double-clause';
+import BaseNumericDoubleClause from '../../components/base/clauses/numeric/base-numeric-double-clause';
 
 export default (props, ClauseComponent) => {
   return () => {
