+++ packages/refine-vue/src/compositions/tailwind/useNumericSingleClause.js
@@ -1,5 +1,5 @@
 import { h } from 'vue-demi';
-import BaseNumericSingleClause from '../../components/tailwind/clauses/numeric/base-numeric-single-clause';
+import BaseNumericSingleClause from '../../components/base/clauses/numeric/base-numeric-single-clause';
 
 export default (props, ClauseComponent) => {
   return () => {
