+++ packages/refine-vue/src/package.js
@@ -2,5 +2,5 @@ import './directives';
 import DatePickerPlugin from './plugins/date-picker';
 import './assets/styles/tailwind.css';
 
-export * from './components/tailwind';
+export * from './components/base';
 export { DatePickerPlugin };
